Fix System.OverflowException for Hsl with double in invariant string (#8114) fixes #8113

+++ Xamarin.Forms.Core/ColorTypeConverter.cs
@@ -251,7 +251,7 @@ namespace Xamarin.Forms
 				maxValue = 100;
 				elem = elem.Substring(0, elem.Length - 1);
 			}
-			return (double)(int.Parse(elem, NumberStyles.Number, CultureInfo.InvariantCulture).Clamp(0, maxValue)) / maxValue;
+			return double.Parse(elem, NumberStyles.Number, CultureInfo.InvariantCulture).Clamp(0, maxValue) / maxValue;
 		}
 
 		static double ParseOpacity(string elem)
